[OPT] Allocate the new item outside the lock.
Since the item allocation can be an expensive operation it's better
to perform it before an item has beed spinlocked, to not make the
spinlock waiters spend their time waiting for the item allocation.

This sounds right, but the performance effect is unmeasurable,
because the existing benchmarks are too unstable.